feat: Create proper results for CustomInterface
First try
