+++ web/libs/editor/src/regions/index.js
@@ -15,7 +15,6 @@ import { HtxTextAreaRegion, TextAreaRegionModel } from "./TextAreaRegion";
 import { RichTextRegionModel } from "./RichTextRegion";
 import { TimelineRegionModel } from "./TimelineRegion";
 import { VideoRectangleRegionModel } from "./VideoRectangleRegion";
-import { CustomRegionModel } from "./CustomRegion";
 
 const AllRegionsType = types.union(
   AudioRegionModel,
@@ -33,7 +32,6 @@ const AllRegionsType = types.union(
   TimelineRegionModel,
   ParagraphsRegionModel,
   VideoRectangleRegionModel,
-  CustomRegionModel,
   ...Registry.customTags.map((t) => t.region).filter(Boolean),
 );
 
@@ -62,5 +60,4 @@ export {
   TextAreaRegionModel,
   TimelineRegionModel,
   VideoRectangleRegionModel,
-  CustomRegionModel,
 };
